+++ v8js_v8.cc
@@ -95,7 +95,7 @@ void v8js_v8_init() /* {{{ */
  * heap allocated memory).
  */
 void v8js_v8_call(v8js_ctx *c, zval **return_value,
-				  long flags, long time_limit, long memory_limit,
+				  long flags, long time_limit, size_t memory_limit,
 				  std::function< v8::Local<v8::Value>(v8::Isolate *) >& v8_call) /* {{{ */
 {
 	char *tz = NULL;
